src/demo/python/requirements.txt
aws-cdk.core
cdk-comprehend-s3olap
